requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
requests==2.31.0
python-magic==0.4.27
Werkzeug==3.0.1
email-validator==2.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
